add withdrawal example with manual checkpoint
